requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
jiter==0.5.0
openai==1.43.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic-core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.5
typing-extensions==4.12.2
